# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aeko Enhanced API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Aeko Development Team",
      "email": "dev@aeko.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "profilePicture": {
            "type": "string",
            "description": "Profile picture URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "away",
              "busy"
            ],
            "description": "User status"
          },
          "botEnabled": {
            "type": "boolean",
            "description": "Whether AI bot auto-reply is enabled"
          },
          "botPersonality": {
            "type": "string",
            "enum": [
              "friendly",
              "professional",
              "sarcastic",
              "creative",
              "analytical",
              "mentor",
              "companion"
            ],
            "description": "Preferred AI bot personality"
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Whether user has admin privileges"
          },
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User's full name"
          },
          "bio": {
            "type": "string",
            "description": "User bio"
          },
          "blueTick": {
            "type": "boolean",
            "description": "Blue tick verification status"
          },
          "goldenTick": {
            "type": "boolean",
            "description": "Golden tick verification status"
          },
          "emailVerification": {
            "type": "object",
            "properties": {
              "isVerified": {
                "type": "boolean"
              }
            }
          },
          "profileCompletion": {
            "type": "object",
            "properties": {
              "completionPercentage": {
                "type": "number"
              },
              "hasProfilePicture": {
                "type": "boolean"
              },
              "hasBio": {
                "type": "boolean"
              },
              "hasFollowers": {
                "type": "boolean"
              },
              "hasVerifiedEmail": {
                "type": "boolean"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "507f1f77bcf86cd799439011",
          "name": "John Doe",
          "username": "johndoe",
          "email": "john@example.com",
          "profilePicture": "https://example.com/profile.jpg",
          "bio": "Tech enthusiast",
          "blueTick": true,
          "goldenTick": false,
          "emailVerification": {
            "isVerified": true
          },
          "profileCompletion": {
            "completionPercentage": 100,
            "hasProfilePicture": true,
            "hasBio": true,
            "hasFollowers": true,
            "hasWalletConnected": true,
            "hasVerifiedEmail": true
          },
          "createdAt": "2024-01-01T12:00:00Z",
          "updatedAt": "2024-01-01T12:00:00Z"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/interests": {
      "post": {
        "summary": "Create a new interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "displayName"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interest created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Interest already exists"
          }
        }
      },
      "get": {
        "summary": "Get all active interests",
        "tags": [
          "Interests"
        ],
        "responses": {
          "200": {
            "description": "List of active interests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/interests/{id}": {
      "put": {
        "summary": "Update an interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interest updated successfully"
          },
          "404": {
            "description": "Interest not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interest deleted successfully"
          },
          "404": {
            "description": "Interest not found"
          }
        }
      }
    },
...
    "/api/user/interests": {
      "get": {
        "summary": "Get user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User's interests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "interestIds"
                ],
                "properties": {
                  "interestIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of interest IDs"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interests updated successfully"
          },
          "400": {
            "description": "Invalid interest IDs"
          }
        }
      }
    },
    "/api/user/interests/{interestId}": {
      "delete": {
        "summary": "Remove an interest from user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "interestId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interest removed successfully"
          },
          "404": {
            "description": "Interest not found in user's interests"
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "68cad398b391bdd7d991d5c7"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of users per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search users by name or username"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "totalUsers": {
                          "type": "integer"
                        },
                        "hasNext": {
                          "type": "boolean"
                        },
                        "hasPrev": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}